ml-server/requirements.txt
flask
flask-cors
python-dotenv
requests
numpy
tensorflow
librosa
scipy
minio
psycopg2-binary
PyJWT==2.8.0
gunicorn
